Version 0.9.0
This is a complete AJAXified version of the app. It uses jQuery to send
off the data to a server and retrieve the resulting, processed, data to
the appropriate part of the page at which point Boostrap kicks in to
return a the modal dialog. Left on the to do for a complete version: a
database interface that stores the values submitted for easy reference
later.